+++ KSol.ZeroVDI/Data/Migrations/20260713134923_AddBackendIdleReapEnabled.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace KSol.ZeroVDI.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddBackendIdleReapEnabled : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "IdleReapEnabled",
+                table: "ProxmoxBackends",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IdleReapEnabled",
+                table: "ProxmoxBackends");
+        }
+    }
+}